implemented the "upkeep loop", which runs every second and handles tracking which users are on break (formerly controlled with timeouts).
each second, it looks at all users that are "on break", and determines if their break has expired. it will switch them to being "over break" if so. it will also handle sending reminders for users that are "over break".

finally, for all users on any sort of break: if you set yourself to "accepting chats" at any point while on break, the bot picks up on that and silently terminates your break.