the bot now checks for and logs bounced chats to the "bounces" table. if a bounce occurred in the last 60 seconds, it'll ping the current chat captain about it.
the bot is also restricted to posting messages in a certain channel. this is checked for when each new message comes in. currently, the only channel it will work in is breakbot_test. this will be expanded so that it'll work when two instances of the bot are running (one in a "dev" channel, one in the chat channel)